created TranslationsAPIClient to handle translations API calls
